Move file status elements to own component
